# swagger2.yaml
openapi: 3.0.2
info:
  title: RAS/RM Party Service
  description: A service for dealing with various parties (businesses, survey respondents etc and their enrolments)
  version: "2.0"
servers:
  - url: https://localhost:8081/party-api/v2
    description: The default path for Party in ras-rm-docker-dev
...
paths:
  /info:
    get:
      summary: Returns basic information about the service.
      description: Returns service information. If the `REPORT_DEPENDENCIES` environment variable is set to true, it will also report information on service dependencies.
      security: []
      tags:
        - info
      servers:
        - url: https://localhost:8081
          description: The /info path doesn't use the /party-api/v1 prefix
      responses:
        '200':
          description: A JSON of basic information about the service.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: ras-rm-party
                  version:
                    type: string
                    example: "1.12.0"
        '404':
          description: The basic information about the service couldn't be retrieved. The service may be configured incorrectly.
  /respondents:
    get:
      summary: Searches for a respondent based on provided attributes.
      description: Searches using the provided attributes using an AND search, returning any respondents that match the criteria.
      tags:
        - respondents
      parameters:
        - in: query
          name: firstName
          schema:
            type: string
            example: Jenny
        - in: query
          name: lastName
          schema:
            type: string
            example: Testerson
        - in: query
          name: emailAddress
          schema:
            type: string
            format: email
            example: test@email.com
        - in: query
          name: telephone
          schema:
            type: string
            example: "01234567890"
        - in: query
          name: status
          schema:
            type: string
            example: SUSPENDED
        - in: query
          name: businessId
          schema:
            type: string
            format: uuid
            example: fd6a1aa3-ba17-43a8-beae-a39e67c6444d
        - in: query
          name: surveyId
          schema:
            type: string
            format: uuid
            example: 71323711-6518-49d5-9767-c42b9fb03639
        - in: query
          name: offset
          schema:
            type: number
            format: integer
            default: 0
        - in: query
          name: limit
          schema:
            type: number
      responses:
        '200':
          description: Respondent(s) were retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RespondentDetails'
        '400':
          $ref: '#/components/responses/QueryParametersMissingError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: No respondents were found based on the parameters provided.
        '500':
          $ref: '#/components/responses/CommunicationError'
    post:
      summary: Creates a new respondent.
      description: | 
        Creates a new respondent from the data in the RequestBody. 
        If an `ID` field is provided, it will use this rather than generating a new UUID.
      tags:
        - respondents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  $ref: '#/components/schemas/RespondentDetails'
                enrolmentCodes:
                  type: array
                  items:
                    type: string
                    example: g023at6a
                    minItems: 1
      responses:
        '201':
          description: The respondent was created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RespondentDetails'
        '400':
          description: A required field (attributes or enrolment code) wasn't provided or was in an incorrect format.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A provided enrolment code couldn't be found.
        '422':
          description: A provided enrolment code has expired or the business associated with an enrolment code couldn't be associated with.
        '500':
          $ref: '#/components/responses/CommunicationError'
  /respondents/{id}:
    delete:
      summary: Deletes a respondent from the service.
      description: | 
        Unenrols a respondent from all surveys then deletes their enrolment history, business associations, pending enrolments and their representation in the service. 
        Will not delete their account from the auth service.
      tags:
        - respondents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
            example: 34597808-ec88-4e93-af2f-228e33ff7946
      responses:
        '204':
          description: The respondent was successfully deleted.
        '400':
          $ref: '#/components/responses/MalformedIDError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/RespondentNotFoundError'
        '500':
          description: There were problems communicating with the database.
    get:
      summary: Retrieves a respondent from the service.
      description: Retrieves a respondent by the ID provided.
      tags:
        - respondents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
            example: 34597808-ec88-4e93-af2f-228e33ff7946
      responses:
        '200':
          description: The respondent was retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RespondentDetails'
        '400':
          $ref: '#/components/responses/MalformedIDError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/RespondentNotFoundError'
    patch:
      summary: Updates the specified fields on the respondent.
      description: | 
        Updates all of the fields specified in the RequestBody for the respondent. 
        This is an atomic operation i.e. if one field can't be updated for whatever reason, none of them will be. 
        Associations will be created if they don't already exist, else they will be updated.
        `ID` is a valid field in the RequestBody, but shouldn't be changed and a `400 Bad Request` will returned if changing it is attempted.
        The respondent will be enrolled on any surveys associated to any provided `enrolmentCodes`.
      tags:
        - respondents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
            example: 34597808-ec88-4e93-af2f-228e33ff7946
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  $ref: '#/components/schemas/RespondentDetails'
                enrolmentCodes:
                  type: array
                  items:
                    type: string
                    example: g023at6a
                    minItems: 0
      responses:
        '200':
          description: The respondent was successfully updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RespondentDetails'
        '400':
          $ref: '#/components/responses/InvalidRequestBodyError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: The respondent wasn't found, one of the associated entities wasn't found by its ID or one of the provided enrolment codes wasn't found.
        '409':
          description: The `emailAddress` provided is already in use for a different respondent or the respondent is already enrolled on a survey attached to a provided enrolment code.
        '422':
          description: Part or all of the update failed, and the action has been rolled back. No enrolments have been generated.
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      headers:
        WWW_Authenticate:
          schema:
            type: string
    MalformedIDError:
      description: The ID provided wasn't a proper UUID.
    RespondentNotFoundError:
      description: The respondent specified wasn't found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Respondent does not exist
...
    CommunicationError:
      description: There were problems communicating with another service or the database.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Couldn't communicate with Case service: 404 not found"
    QueryParametersMissingError:
      description: No query parameters were provided.
    InvalidRequestBodyError:
      description: One or more of the fields provided in the RequestBody wasn't part of the schema, wasn't set to a valid value, or the ID provided wasn't a proper UUID.
  schemas:
    RespondentDetails:
      type: object
      properties:
        attributes:
          type: object
          properties:
            emailAddress:
              type: string
              format: email
              example: test@email.com
            id:
              type: string
              format: uuid
              example: 34597808-ec88-4e93-af2f-228e33ff7946
            lastName:
              type: string
              example: Testerson
            firstName:
              type: string
              example: Jenny
            telephone:
              type: string
              example: "01234567890"
        status:
          type: string
          example: ACTIVE
        associations:
          type: array
          items:
            type: object
            properties:
              enrolments:
                type: array
                items:
                  type: object
                  properties:
                    enrolmentStatus:
                      type: string
                      example: ENABLED
                    surveyId:
                      type: string
                      format: uuid
                      example: 71323711-6518-49d5-9767-c42b9fb03639
              id: 
                type: string
                format: uuid
                example: fd6a1aa3-ba17-43a8-beae-a39e67c6444d
...
security:
  - basicAuth: []